refactor(app-builder-sdk): examples/ → reference/<shape>-<provider>/ + clarify role
Renames to remove the "Slack v2" / "demo vs production" ambiguity. Each
reference app's directory name now states the SDK shape it demonstrates:

  examples/pinterest        → reference/pinterest-publishing
  examples/slack            → reference/slack-messaging
  examples/github-issues    → reference/github-workflow
  examples/gcalendar        → reference/gcalendar-events
  examples/telegram         → reference/telegram-messaging

Each app.ts gains a top-of-file banner that says explicitly:
  - "SDK REFERENCE — NOT a production module"
  - what shape / concepts it demonstrates
  - the relationship to hola-boss-apps/<name>/ (the legacy production
    module of the same provider, if any) — e.g. reference/slack-messaging
    has a different / narrower tool surface than hola-boss-apps/slack/;
    one does not replace the other
  - guidance to copy as a template rather than deploy verbatim

README's "What's the SDK, what isn't" rewritten to make explicit the
three-place model:
  - experiments/app-builder-sdk/reference/<shape>/ → SDK validation +
    test fixtures + agent templates (this repo)
  - hola-boss-apps/<name>/ → legacy production modules built on the old
    @holaboss/bridge SDK (~20 of them; sunset is a multi-week migration
    tracked separately)
  - <workspace>/apps/<id>/ → workspace-installed apps materialized via
    marketplace install or the app-builder skill

What didn't change: SDK behavior, public API, test surface (82 pass /
343 expect / tsc clean — same as before rename).
