+++ experiments/app-builder-sdk/test/github-issues.test.ts
@@ -1,7 +1,7 @@
 import { describe, test, expect } from "bun:test"
 import { createBridge, type TransportFn } from "../src/bridge.ts"
-import { GITHUB } from "../examples/github-issues/provider.ts"
-import { buildGithubIssuesApp } from "../examples/github-issues/app.ts"
+import { GITHUB } from "../reference/github-workflow/provider.ts"
+import { buildGithubIssuesApp } from "../reference/github-workflow/app.ts"
 import type { AppHandleInternal } from "../src/app.ts"
 
 let calls: Array<{ method: string; url: string; body?: any }> = []
