+++ experiments/app-builder-sdk/test/telegram.test.ts
@@ -1,7 +1,7 @@
 import { describe, test, expect } from "bun:test"
 import { createBridge, type TransportFn } from "../src/bridge.ts"
-import { TELEGRAM } from "../examples/telegram/provider.ts"
-import { buildTelegramApp } from "../examples/telegram/app.ts"
+import { TELEGRAM } from "../reference/telegram-messaging/provider.ts"
+import { buildTelegramApp } from "../reference/telegram-messaging/app.ts"
 import type { AppHandleInternal } from "../src/app.ts"
 
 let calls: Array<{ method: string; url: string; body?: any }> = []
